Fix build after API changes to QColor

+++ src/quick/util/qquickvaluetypes.cpp
@@ -148,42 +148,42 @@ void QQuickColorValueType::setA(qreal a)
 
 void QQuickColorValueType::setHsvHue(qreal hsvHue)
 {
-    qreal hue, saturation, value, alpha;
+    float hue, saturation, value, alpha;
     v.getHsvF(&hue, &saturation, &value, &alpha);
     v.setHsvF(hsvHue, saturation, value, alpha);
 }
 
 void QQuickColorValueType::setHsvSaturation(qreal hsvSaturation)
 {
-    qreal hue, saturation, value, alpha;
+    float hue, saturation, value, alpha;
     v.getHsvF(&hue, &saturation, &value, &alpha);
     v.setHsvF(hue, hsvSaturation, value, alpha);
 }
 
 void QQuickColorValueType::setHsvValue(qreal hsvValue)
 {
-    qreal hue, saturation, value, alpha;
+    float hue, saturation, value, alpha;
     v.getHsvF(&hue, &saturation, &value, &alpha);
     v.setHsvF(hue, saturation, hsvValue, alpha);
 }
 
 void QQuickColorValueType::setHslHue(qreal hslHue)
 {
-    qreal hue, saturation, lightness, alpha;
+    float hue, saturation, lightness, alpha;
     v.getHslF(&hue, &saturation, &lightness, &alpha);
     v.setHslF(hslHue, saturation, lightness, alpha);
 }
 
 void QQuickColorValueType::setHslSaturation(qreal hslSaturation)
 {
-    qreal hue, saturation, lightness, alpha;
+    float hue, saturation, lightness, alpha;
     v.getHslF(&hue, &saturation, &lightness, &alpha);
     v.setHslF(hue, hslSaturation, lightness, alpha);
 }
 
 void QQuickColorValueType::setHslLightness(qreal hslLightness)
 {
-    qreal hue, saturation, lightness, alpha;
+    float hue, saturation, lightness, alpha;
     v.getHslF(&hue, &saturation, &lightness, &alpha);
     v.setHslF(hue, saturation, hslLightness, alpha);
 }
